4-2_api_basics/02_kadai_4_2_2_08_API2/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.4
google-api-python-client==2.108.0
pytz==2023.3
